fix: cast mmsi and imo_number in postgres static insert
Cast mmsi and imo_number to i32 when binding the postgres static
insert in db.rs so the values match the INTEGER columns. Also import
min alongside max in src/lib.rs to simplify the worker count
calculation.

+++ aisdb_lib/src/db.rs
@@ -176,12 +176,12 @@ pub fn postgres_insert_static(
         tx.execute(
             &stmt,
             &[
-                &p.mmsi,
+                &(p.mmsi as i32),
                 &(e as i32),
                 &p.name.unwrap_or_default(),
                 &(p.ship_type as i32),
                 &p.call_sign.unwrap_or_default(),
-                &p.imo_number.unwrap_or_default(),
+                &(p.imo_number.unwrap_or_default() as i32),
                 &(p.dimension_to_bow.unwrap_or_default() as i32),
                 &(p.dimension_to_stern.unwrap_or_default() as i32),
                 &(p.dimension_to_port.unwrap_or_default() as i32),
